harbour and vertex setup moved from Satan into Board
