+++ classes/Satan.js
@@ -1,9 +1,5 @@
 const Board = require("./Board.js")
-const Hex = require("./Hex.js")
-const Vertex = require("./Vertex.js")
 const Edge = require("./Edge.js")
-const Graph = require("./Graph.js")
-const Harbour = require("./Harbour.js")
 const DevelopmentCard = require("./DevelopmentCard.js")
 
 const lobbies = require("../server/lobbies.js")
@@ -85,7 +81,7 @@ class Satan {
     }
     
     const tickData = {
-      board: this.board.hexes,
+      board: this.board.hexes, // legacy data format (by which i mean i dont want to rewrite the frontend right now)
       vertexes: this.board.vertexes,
       edges: this.board.edges,
       players: playersPublicData,
@@ -120,10 +116,7 @@ class Satan {
     else this.players[playerId] = data
   }
 
-  getVertex(coords) {
-    if(coords == undefined || coords.x == undefined || coords.y == undefined || coords.v == undefined) return null
-    return this.board.vertexes.filter(e => e.coords.x === coords.x && e.coords.y === coords.y && e.coords.v === coords.v)[0]
-  }
+  getVertex(coords) { return this.board.getVertex(coords) }
 
   getEdge(coordsArr) { // i dont know how efficient this is but it does work without requiring the coords to be in a specific order
     return this.board.edges.filter(e => {
@@ -152,86 +145,6 @@ class Satan {
     this.board = new Board()
 
     this.board.setup(playerCount)
-
-
-    let harbourTypeCounts = {
-      ore: 1,
-      wheat: 1,
-      bricks: 1,
-      wool: 1,
-      lumber: 1,
-      any: 4
-    }
-    
-    for(let row of this.board.hexes) { // calculate harbours
-      for(let hex of row) {
-        if(!hex?.invisible) continue
-
-        const adjHexCoords = hex.getAdjacentHexes()
-        const adjHexes = []
-        adjHexCoords.forEach(c => {
-          const loopHex = this.board.getRow(c.y)?.[c.x]
-          if(loopHex) adjHexes.push(loopHex)
-        })
-
-        if(adjHexes.every(adjHex => !adjHex.harbour)) {
-          const { x, y } = hex.coords
-          const hexQuadrant = {
-            x: (x >= this.board.hexes[0].length / 2), // values are true if in positive quadrant, false otherwise
-            y: (y >= this.board.hexes.length / 2), // also the positive y quadrant is the bottom one i will definitely forget that if i dont comment this
-          }
-
-          const adjVertCoords = hex.getAdjacentVertexes()
-          const adjVerts = []
-          adjVertCoords.forEach(c => {
-            const loopVert = this.getVertex(c)
-            if(loopVert) adjVerts.push(loopVert)
-          })
-
-          let harbourVerts = []
-          if(adjVerts.length <= 2) { // if there are already only two verts adjacent
-            harbourVerts = adjVerts // just use those two and skip the logic in the else below
-          }
-          else {
-            if(!hexQuadrant.x) { // logic for left side of board
-              harbourVerts.push(
-                adjVerts.filter(vert => vert.coords.y === (hexQuadrant.y ? y       : y - 1  ))[0],
-                adjVerts.filter(vert => vert.coords.v === (hexQuadrant.y ? "south" : "north"))[0],
-              )
-            }
-            else {
-              // logic for right side (very nice variable names ahead :D)
-              // all of this will probably break with any other board layout which would be fun
-
-              const specialCase1 = adjVerts.filter(vert => vert.coords.y === y + 1)
-              const specialCase1B = adjVerts.filter(vert => vert.coords.y === y) // special case for the bottom harbour that messes everything up
-              harbourVerts.push(specialCase1[0] || specialCase1B[0])
-              
-              const specialCase2 = adjVerts // handles the harbours on the top and bottom edges
-                .filter(vert => vert.coords.y === (hexQuadrant.y ? y - 1 : y))
-                .sort((a, b) => a.coords.x - b.coords.x)
-              harbourVerts.push(specialCase2[0])
-            }
-          }
-
-
-
-
-          const finalHarbourVerts = harbourVerts.filter(v => v) // find all vertexes that are defined
-          finalHarbourVerts.forEach(vert => vert.harbour = hex.coords) // define in the vertexes the coords of the harbour hex
-
-
-          const harbour = new Harbour(finalHarbourVerts.map(v => v.coords))
-
-          const possibleHarbours = Object.keys(harbourTypeCounts).filter(k => harbourTypeCounts[k] > 0)
-          const harbourType = possibleHarbours[Math.floor(Math.random() * possibleHarbours.length)]
-          harbourTypeCounts[harbourType]--
-          harbour.setDeal(harbourType, harbourType === "any" ? 3 : 2)
-
-          hex.setHarbour(harbour)
-        }
-      }
-    }
 
     for(let row of this.board.hexes){
       for(let hex of row){
